refactor: フォルダの管理方法の変更、use_DBをcreate_DBに変更
