+++ pkg/cruds/read/get_hint_list.go
