+++ pkg/cruds/read/is_mode_wolf.go
