+++ pkg/cruds/read/is_room_exit.go
