+++ pkg/cruds/read/result.go
